Can now change to fullscreen during runtime
	modified:   include/Axios/Instance.h
	modified:   src/Axios/Instance.cpp

The Instance class now serves as a singleton class.
This is to allow the window to be changed to fullscreen during runtime.
SFML doesn't support this, so this is a work around. It also allows us
to change the fps limit once without converting it all the time from
a string to a integer.