Scenes now serialize objects
	modified:   include/Axios/Object.h
	modified:   include/Axios/Scene.h
	modified:   src/Axios/Instance.cpp
	modified:   src/Axios/Object.cpp
	modified:   src/Axios/Scene.cpp
	modified:   test/Main.cpp

Serialiation of object is not that flexible as of now. Work
should be invested in save location of scenes, maybe some kind
of generic location.
Either way, objects now save their state. If a save file for a scene
exists, that save file is loaded. right now, there is no way to
change that.